shell: Add a new IconTheme provider for image:// URLs in QML
This will allow us to access images from the icon theme within our QML
files to allow for some proper integration. Additionally, we ensure we
properly share the QQmlEngine instance across all of the shell for now
so that the right provider is shared globally.

We will eventually add symbolic icon theme support to this provider likely
using masks or similar.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>