session: If XDG_CURRENT_DESKTOP is used, we must respect it
If the XDG_CURRENT_DESKTOP variable is set in the environment, we'll respect
that value and use it internally. This not only makes sure we do the right
thing, but actually allows us to test the components under other desktops
to make sure stuff is working OK.

If the variable is not set, i.e. because we've been set through some janky
X11 startup script, we'll forcibly set the variable.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>