panel: Begin skeleton for the new MenuWindow
The MenuWindow will be the main container for the Budgie Menu. Initially
this is going to be a bit of a demo (read: utter shite) much like the Vala
variant in Budgie 10.

However, once we've got our base widgets in (such as Budgie::Popover), we'll
gut this cruft out and take a modular approach to large components such as
the menu, and share as much as the inner "guts" with the rest of the codebase
as we can. This will allow us to take an approach like Brisk, but make sure
that the internals are shared library-wide for other applet developers too.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>