session: Forcibly override DESKTOP_SESSION/XDG_DESKTOP_NAME
We must ensure these are the correct values, as we know from the past we
had numerous issues in display managers deciding to alter the name as they
saw fit.

Should we need to provide compatibility in future, we can use a conjoined
desktop name within the session, but ensure to disable autostart for the
incompatible applications.

It is likely we'll set "Budgie:GNOME" again in future so that applications
"fit in", hence this override approach.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>