shell: Add new start() routine
We'll do this from within the main loop now as we're going to try loading
plugins that require a working dispatch system. We'll set this as the first
application event using a singleShot timer and attempt start within the
main loop.

Right now this deliberately fails as we have no way to load the notifications
plugin.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>