faces: Create new FaceInterface for defining UI
The FaceInterface allows us to define and implement the main UI portion,
and the default Face will implement a similar experience to Budgie 10,
but obviously using Qt.

The Face is responsible for all graphical operations, i.e. panels, OSDs,
etc. It should not implement window management however, this is a reserved
operation for the windowmanager service plugin.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>