Add new Raven interface and fix existing IIDs
This will allow the shell Face/Plugins to easily access any Raven instance,
if one is currently registered as running. The idea will be that we actually
implement the main Raven implementation as a shared library and have the
DefaultFace implementation register a default Raven instance. However this
will allow future Face implementations to have their own "Raven" and register
it in a way that existing widgets will be able to talk to it in a standard
fashion.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>