Start a new session component
Right now the shell entry is more complex that it needs to be, which in
turn is breaking event loops. Instead we'll split the main session into
an orchestration binary to bootstrap the environment and launch both the
window manager and shell. This will fix the chicken & egg issues we have
right now with the shell implementation.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>