Rename Window to SurfaceItem
We need to start separating a Surface from a Window, as the display is
required to compose a window (xdg/wl_shell) of surfaces, but the manager
must know about all of those surfaces.

For now we're doing a simple mass-replace, but we'll start refining this
and building Window objects from these surfaces to provide to the renderer.
Effectively we'll not allow any surface that is missing a window, and those
surfaces will remain unmapped unless we can specialise them into something
like a cursor surface.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>