Not yet working as we map too soon, but preemptively set cursor layer
Right now we're mapping everything directly to screen as soon as we know
about the surface, which is why the render order is still borked. Instead
we're going to have to separate the notion of surface items and windows,
and compose windows surfaces. As and when the surface is ready for mapping,
i.e. it has content, we'll then map a window for it.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>