Add framework for WaylandWindow promotion
This will need backend changes before we can make use of it, but we're
slowly converting our reliance on a root level surface into a lazily
created window as and when we have a way to contain it. For now we're
just gonna set the routine up to construct a new Window.

In future we'll **only** render the WaylandWindow surfaces and we won't
add arbitrary surfaces to the render loop.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>