Teach the compositor to render windows, not SurfaceItems
OK so technically we're still rendering SurfaceItems as part of a Window
but this does ensure we have proper parenting and greatly minimises the
amount of work our renderer needs to do, by only concerning itself with
drawing and checking visible surfaces.

Prior to this we'd find ourselves checking input and rendering on surfaces
such as cursors, for every single application.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>