Fix some performance issues
For now we're going to only set the vsync on the first window and allow all
windows to be governed by the same refresh rate. The reality is we're probably
going to need to add our own vsync-esque approach and separate rendering loops
for each window in time. For now, this'll fix our performance issues.

Additionally we bind the root of each window to an FBO and allow partial
blits to improve overall percieved performance of animations, and ensure
that our window has at minimum a double buffer.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>